EVAKA-TECH: use action based access control - part 3/x

+++ service/src/main/kotlin/fi/espoo/evaka/placement/PlacementController.kt
@@ -8,8 +8,6 @@ import fi.espoo.evaka.Audit
 import fi.espoo.evaka.EvakaEnv
 import fi.espoo.evaka.daycare.controllers.AdditionalInformation
 import fi.espoo.evaka.daycare.controllers.Child
-import fi.espoo.evaka.daycare.controllers.utils.noContent
-import fi.espoo.evaka.daycare.controllers.utils.ok
 import fi.espoo.evaka.daycare.createChild
 import fi.espoo.evaka.daycare.getChild
 import fi.espoo.evaka.shared.DaycareId
@@ -21,36 +19,33 @@ import fi.espoo.evaka.shared.async.GenerateFinanceDecisions
 import fi.espoo.evaka.shared.auth.AccessControlList
 import fi.espoo.evaka.shared.auth.AclAuthorization
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
-import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.BadRequest
 import fi.espoo.evaka.shared.domain.DateRange
-import fi.espoo.evaka.shared.domain.NotFound
+import fi.espoo.evaka.shared.security.AccessControl
+import fi.espoo.evaka.shared.security.Action
 import org.springframework.format.annotation.DateTimeFormat
-import org.springframework.http.ResponseEntity
-import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
-import java.net.URI
+import org.springframework.web.bind.annotation.RestController
 import java.time.LocalDate
 import java.util.UUID
 
-@Controller
-@RequestMapping("/placements")
+@RestController
 class PlacementController(
     private val acl: AccessControlList,
+    private val accessControl: AccessControl,
     private val asyncJobRunner: AsyncJobRunner,
     env: EvakaEnv
 ) {
     private val useFiveYearsOldDaycare = env.fiveYearsOldDaycareEnabled
 
-    @GetMapping
+    @GetMapping("/placements")
     fun getPlacements(
         db: Database.Connection,
         user: AuthenticatedUser,
@@ -64,28 +59,28 @@ class PlacementController(
             value = "to",
             required = false
         ) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) endDate: LocalDate? = null
-    ): ResponseEntity<Set<DaycarePlacementWithDetails>> {
+    ): Set<DaycarePlacementWithDetails> {
         Audit.PlacementSearch.log(targetId = daycareId ?: childId)
-
-        val roles = when {
-            daycareId != null -> acl.getRolesForUnit(user, daycareId)
-            childId != null -> acl.getRolesForChild(user, childId)
+        when {
+            daycareId != null -> accessControl.requirePermissionFor(user, Action.Unit.READ_PLACEMENT, daycareId)
+            childId != null -> accessControl.requirePermissionFor(user, Action.Child.READ_PLACEMENT, childId)
             else -> throw BadRequest("daycareId or childId is required")
         }
-        roles.requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF, UserRole.SPECIAL_EDUCATION_TEACHER)
+
         val auth = acl.getAuthorizedDaycares(user)
         val authorizedDaycares = auth.ids ?: emptySet()
 
         return db.read {
             it.getDetailedDaycarePlacements(daycareId, childId, startDate, endDate).map { placement ->
+                // TODO: is some info only hidden on frontend?
                 if (auth !is AclAuthorization.All && !authorizedDaycares.contains(placement.daycare.id))
                     placement.copy(isRestrictedFromUser = true)
                 else placement
-            }.toSet().let(::ok)
+            }.toSet()
         }
     }
 
-    @GetMapping("/plans")
+    @GetMapping("/placements/plans")
     fun getPlacementPlans(
         db: Database.Connection,
         user: AuthenticatedUser,
@@ -95,23 +90,21 @@ class PlacementController(
             required = false
         ) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) startDate: LocalDate,
         @RequestParam(value = "to", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) endDate: LocalDate
-    ): ResponseEntity<List<PlacementPlanDetails>> {
+    ): List<PlacementPlanDetails> {
         Audit.PlacementPlanSearch.log(targetId = daycareId)
-        acl.getRolesForUnit(user, daycareId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Unit.READ_PLACEMENT_PLAN, daycareId)
 
-        return db.read { it.getPlacementPlans(daycareId, startDate, endDate) }.let(::ok)
+        return db.read { it.getPlacementPlans(daycareId, startDate, endDate) }
     }
 
-    @PostMapping
+    @PostMapping("/placements")
     fun createPlacement(
         db: Database.Connection,
         user: AuthenticatedUser,
         @RequestBody body: PlacementCreateRequestBody
-    ): ResponseEntity<Unit> {
+    ) {
         Audit.PlacementCreate.log(targetId = body.childId, objectId = body.unitId)
-        acl.getRolesForUnit(user, body.unitId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Unit.CREATE_PLACEMENT, body.unitId)
 
         if (body.startDate > body.endDate) throw BadRequest("Placement start date cannot be after the end date")
 
@@ -141,19 +134,17 @@ class PlacementController(
         }
 
         asyncJobRunner.scheduleImmediateRun()
-        return ResponseEntity.noContent().build()
     }
 
-    @PutMapping("/{placementId}")
+    @PutMapping("/placements/{placementId}")
     fun updatePlacementById(
         db: Database.Connection,
         user: AuthenticatedUser,
         @PathVariable("placementId") placementId: PlacementId,
         @RequestBody body: PlacementUpdateRequestBody
-    ): ResponseEntity<Unit> {
+    ) {
         Audit.PlacementUpdate.log(targetId = placementId)
-        acl.getRolesForPlacement(user, placementId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Placement.UPDATE, placementId)
 
         val aclAuth = acl.getAuthorizedDaycares(user)
         db.transaction { tx ->
@@ -173,18 +164,16 @@ class PlacementController(
         }
 
         asyncJobRunner.scheduleImmediateRun()
-        return ResponseEntity.noContent().build()
     }
 
-    @DeleteMapping("/{placementId}")
+    @DeleteMapping("/placements/{placementId}")
     fun deletePlacement(
         db: Database.Connection,
         user: AuthenticatedUser,
         @PathVariable("placementId") placementId: PlacementId
-    ): ResponseEntity<Unit> {
+    ) {
         Audit.PlacementCancel.log(targetId = placementId)
-        acl.getRolesForPlacement(user, placementId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Placement.DELETE, placementId)
 
         db.transaction { tx ->
             val (childId, startDate, endDate) = tx.cancelPlacement(placementId)
@@ -195,19 +184,17 @@ class PlacementController(
         }
 
         asyncJobRunner.scheduleImmediateRun()
-        return ResponseEntity.noContent().build()
     }
 
-    @PostMapping("/{placementId}/group-placements")
+    @PostMapping("/placements/{placementId}/group-placements")
     fun createGroupPlacement(
         db: Database.Connection,
         user: AuthenticatedUser,
         @PathVariable("placementId") placementId: PlacementId,
         @RequestBody body: GroupPlacementRequestBody
-    ): ResponseEntity<GroupPlacementId> {
+    ): GroupPlacementId {
         Audit.DaycareGroupPlacementCreate.log(targetId = placementId, objectId = body.groupId)
-        acl.getRolesForPlacement(user, placementId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Placement.CREATE_GROUP_PLACEMENT, placementId)
 
         return db.transaction { tx ->
             tx.checkAndCreateGroupPlacement(
@@ -215,42 +202,65 @@ class PlacementController(
                 groupId = body.groupId,
                 startDate = body.startDate,
                 endDate = body.endDate
-            ).let {
-                ResponseEntity.created(URI.create("/placements/$placementId/group-placements/$it")).body(it)
-            }
+            )
         }
     }
 
-    @DeleteMapping("/{daycarePlacementId}/group-placements/{groupPlacementId}")
+    @DeleteMapping("/group-placements/{groupPlacementId}")
     fun deleteGroupPlacement(
+        db: Database.Connection,
+        user: AuthenticatedUser,
+        @PathVariable("groupPlacementId") groupPlacementId: GroupPlacementId
+    ) {
+        Audit.DaycareGroupPlacementDelete.log(targetId = groupPlacementId)
+        accessControl.requirePermissionFor(user, Action.GroupPlacement.DELETE, groupPlacementId)
+
+        db.transaction { it.deleteGroupPlacement(groupPlacementId) }
+    }
+
+    // TODO: remove after deployed once
+    @DeleteMapping("/placements/{daycarePlacementId}/group-placements/{groupPlacementId}")
+    fun deleteGroupPlacementDeprecated(
         db: Database.Connection,
         user: AuthenticatedUser,
         @PathVariable("daycarePlacementId") daycarePlacementId: PlacementId,
         @PathVariable("groupPlacementId") groupPlacementId: GroupPlacementId
-    ): ResponseEntity<Unit> {
+    ) {
         Audit.DaycareGroupPlacementDelete.log(targetId = groupPlacementId)
-        acl.getRolesForPlacement(user, daycarePlacementId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.GroupPlacement.DELETE, groupPlacementId)
 
-        val success = db.transaction { it.deleteGroupPlacement(groupPlacementId) }
-        if (!success) throw NotFound("Group placement not found")
-        return noContent()
+        db.transaction { it.deleteGroupPlacement(groupPlacementId) }
     }
 
-    @PostMapping("/{daycarePlacementId}/group-placements/{groupPlacementId}/transfer")
+    @PostMapping("/group-placements/{groupPlacementId}/transfer")
     fun transferGroupPlacement(
+        db: Database.Connection,
+        user: AuthenticatedUser,
+        @PathVariable("groupPlacementId") groupPlacementId: GroupPlacementId,
+        @RequestBody body: GroupTransferRequestBody
+    ) {
+        Audit.DaycareGroupPlacementTransfer.log(targetId = groupPlacementId)
+        accessControl.requirePermissionFor(user, Action.GroupPlacement.UPDATE, groupPlacementId)
+
+        db.transaction {
+            it.transferGroup(groupPlacementId, body.groupId, body.startDate)
+        }
+    }
+
+    // TODO: remove after deployed once
+    @PostMapping("/placements/{daycarePlacementId}/group-placements/{groupPlacementId}/transfer")
+    fun transferGroupPlacementDeprecated(
         db: Database.Connection,
         user: AuthenticatedUser,
         @PathVariable("daycarePlacementId") daycarePlacementId: PlacementId,
         @PathVariable("groupPlacementId") groupPlacementId: GroupPlacementId,
         @RequestBody body: GroupTransferRequestBody
-    ): ResponseEntity<Unit> {
+    ) {
         Audit.DaycareGroupPlacementTransfer.log(targetId = groupPlacementId)
-        acl.getRolesForPlacement(user, daycarePlacementId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Placement.CREATE_GROUP_PLACEMENT, daycarePlacementId)
 
-        return db.transaction {
-            it.transferGroup(daycarePlacementId, groupPlacementId, body.groupId, body.startDate).let(::noContent)
+        db.transaction {
+            it.transferGroup(groupPlacementId, body.groupId, body.startDate)
         }
     }
 }
